+++ scripts/redis.sh
@@ -64,6 +64,11 @@ if [ $(curl -s -o /dev/null -w "%{http_code}" http://john:doe@127.0.0.1:9990) !=
 	exit 1
 fi
 
+if [ $(curl -s -o /dev/null -w "%{http_code}" http://john:doe@127.0.0.1:9990/healthz) != "200" ]; then
+    echo "Should be OK"
+    exit 1
+fi
+
 if [ $(curl -s -o /dev/null -w "%{http_code}" http://john:doe@127.0.0.1:9990/api/v1/state/redis-2/my-redis) != "200" ]; then
 	echo "Failed to query state"
 	exit 1
